@@ -1,37 +1,64 @@
 <?php
 declare(strict_types = 1);  
-//session_start();                                                      //start or renew session
+use Liquimet\Controller\UserController;
+use Liquimet\Session\Session;
+use Liquimet\Model\User;
+use Twig\Environment;
+use Twig\Loader\FilesystemLoader;
 
-    /*if(!$id){                                                           //if no valid id
-        include APP_ROOT . '/src/views/page-not-found.php';             //page not found
-    }
+// Setup Twig environment
+$twig = new Environment(new FilesystemLoader(APP_ROOT . '/templates'));
 
-    $user = $model->getUser()->get($id);                                //get member data
-        if(!$user){                                                     //if array is empty
-            include APP_ROOT . '/src/views/page-not-found.php';         //page not found
-        }*/
+// Instantiate the session and user model
+$session = new Session(); 
+$userModel = new User( $database);  
+
+// Instantiate and handle login actions
+$userController = new UserController($twig, $session, $userModel);
+
+// Call the index method to load the admin dashboard
+$userController->indexUser();
+
+/*
+$session->requireLogin();
+
+$user = $model->getUser()->get($session->getID());
 
 //navigation menu data 
-$data['navigation'] = $model->getPlatform()->getGroups(); 
-$data['categories'] = $model->getPlatform()->getAll();  
-$data['group_count'] = $model->getPlatform()->countG();  
+$data['navigation'] = $model->getPlatform()->getNavigation(); 
+$data['totalAdminNav'] = $model->getPlatform()->totalAdminNav();  
+
+//login and session data
+$data['user'] = $user;  
 
 //dashboard data
-$data['countU'] = $model->getUser()->count();
+$data['unloadedSum'] = $model->getQuantity()->sumUnloaded();
+
+$data['allTransports'] = $model->getTransport()->totalTransport();
+$data['fullTransports'] = $model->getTransport()->totalTransports('F');
+$data['partTransports'] = $model->getTransport()->totalTransports('P');
+$data['recent'] = $model->getTransport()->getRecent($user['id_user']);
+
+echo $twig->render('index.twig', $data);
+
+//$logged_in = $model->getSession()->username ?? false;
+//require_login($logged_in);
+/*require_login();
+
+//navigation menu data 
+$data['navigation'] = $model->getPlatform()->getAdminGroups(); 
+$data['group_count'] = $model->getPlatform()->countG();  
+
+//login and session data
+$data['user'] = $user = $model->getUser()->get($model->getSession()->id_user);;     
+
+//dashboard data
+$data['countU'] = $model->getUser()->countActive();
 $data['countT'] = $model->getTransport()->count();
 $data['sum'] = $model->getQuantity()->sum();
 $data['countFull'] = $model->getTransport()->countFull();
 $data['countPart'] = $model->getTransport()->countPart();
-$data['recent'] = $model->getTransport()->getRecent();
-$data['admins'] = $model->getUser()->countAdmin();
-$data['members'] = $model->getUser()->countMember();
-$data['offlines'] = $model->getUser()->countSuspended();
+$data['recent'] = $model->getTransport()->getRecent($user['id_user']);
+//$data['trans'] = $model->getTransport()->getAll($user['id_user'], 7);       //get user's transports
 
-//other data
-$data['user'] = $user = $model->getUser()->get($id);                                                                  //user data
-$data['trans'] = $model->getTransport()->getAll($user['id_user'], 5);                   //get user's transports
-
-$data['group'] = '';                                                                    //group id for link
-//$data['section'] = '';    
-
-echo $twig->render('index.html', $data);                       
+echo $twig->render('index.html', $data);       */                   
