@@ -1,38 +1,26 @@
 <?php
-declare(strict_types = 1);   
+declare(strict_types = 1);  
+use Liquimet\Controller\PlatformController;
+use Liquimet\Session\Session;
+use Liquimet\Model\Navigation;
+use Liquimet\Model\User;
+use Liquimet\Model\Transport;
+use Liquimet\Model\Quantity;
+use Twig\Environment;
+use Twig\Loader\FilesystemLoader;
 
-    if(!$id){                                                   
-        include APP_ROOT . '/src/views/page-not-found.php';    
-    }
+// Setup Twig environment
+$twig = new Environment(new FilesystemLoader(APP_ROOT . '/templates'));
 
-$category = $model->getPlatform()->get($id);                 
-    if(!$category){                                          
-        include APP_ROOT . '/src/views/page-not-found.php';    
-    } 
+// Instantiate the session and user model
+$session = new Session(); 
+$navModel = new Navigation($database); 
+$userModel = new User($database);
+$transModel = new Transport($database);
+$qtyModel = new Quantity($database);
 
-    if(mb_strtolower($parts[2]) != mb_strtolower($category['seo_group'])) {                  //if seo name wrong
-        redirect('platform/' . $id . '/' . $category['seo_group'], [], 301);                 //redirect to correct url
-    }   
-        
-//$user = $model->getUser()->get($id);
-     
-//navigation menu data
-$data['navigation'] = $model->getPlatform()->getGroups(); 
-$data['categories'] = $model->getPlatform()->getAll();   
-$data['group_count'] = $model->getPlatform()->countG();  
+// Instantiate and handle login actions
+$platformController = new PlatformController($twig, $session, $navModel, $userModel, $transModel, $qtyModel);
 
-//login and session data
-$data['user'] = $user = $model->getUser()->get($id); 
-
-//table data
-$data['transports'] = $model->getTransport()->getAll();
-$data['quantities'] = $model->getQuantity()->getAll();
-$data['partials'] = $model->getPartial()->getAll();
-$data['total'] = $model->getTransport()->count();
-
-//current category data
-$data['category'] = $category;                                                        //current group
-$data['group'] = $category['id_group'];                                               //group id for link     
-$data['section'] = '';                                                                //category id for nav (on next page of categories) 
-
-echo $twig->render('platform.html', $data);                                         
+// Call the index method to load the admin dashboard
+$platformController->allTransportsPage();
